fix(api): improve reliability of saves/pushes from HK to carrier API
This batches changes to send to the api and hopefully helps with edge
cases where two changes made too close to each other can overwrite one
another. Hopefully helping with #147

It also makes the api more resiliant to how the caller calls the
methods, so we don't have to pass all zone args each time. This will
allow me to finish the refactor to unblock #111