fix(api): fix index vs zone id error in zone enumeration
I've never seen this happen, but if a system had zone 1 disabled, and zone 2 used (or just in
general had non-contiguous zones being used), the accessories created would not actually reference
the right zones. It's possible this could be related to some of the weird, hard to debug issues that
have been reported.