feat(thermostat): support turning on/off manual holds from HK
Adds a manual hold switch with the other activity switches. This indicates that a hold of some kind
is active. This could be an activity hold, or the 'manual' psudo-activity. To match thermostat
behavior, you can't switch directly tothe manual activity. Instead activating this switch turns on a
hold for the rrent activity. To hold to the manual activity, just change the temp.

fix #245