feat: rewrite api to be observer based and poll/push updates to HomeKit (#442)
This is a significant rewrite of the base carrier api away from promises and to observables. This allows the api data model to poll carrier for updates, and for those updates to automatically be pushed to HK.

This also changes the data push model to use the same push api, which simplifiers and hopefully makes it easier to remove edge cases in the set apis.

In my testing, these changes significantly improve the perceived responsiveness of the plugin.